Available to budget can now go negative.
This is the cleanest way to solve catch 22 issues between accounts and budgets when getting started.  It also allows the past to be what it was without tons of errors that are tedious to fix for no benefit.  It also allows users who are just getting started or 'floating' to get useful results without a bunch of warnings.  The end result is the same, and CP displays a warning if you are overbudgeted anyway.