Changed some fatal errors to warnings that we can proceed past
Want to eliminate the 'die' concept so that we can gracefully deal with errors for the sake of a gui future.  There can still be fatal issues reading the input, but all of that can be considered 'read errors' with the input file, before we start trying to make sense of the input data.  Right now the concepts are mixed together.