Resolve issue #3
Account closure is tracked as a flag on the account that doesn't
interfere with the rest of the account data, so closure is really a flag
to say that the account shouldn't be displayed if it has all zero
balances.  Accounts still need to be closed in order to change budget
type.