Implemented #9
When a new budget comes into effect, all the existing categories are left alone.  In order to change a category, or remove it, a new category type of "close" can be used to remove the category, and then the category can be created freshly.  If the category had money in it, CP returns it to the category owner's available and reports it at runtime.