Implemented #2
Budgets can now contain a 'withholding' category type that is always autofunded from its owner's available with no warnings or other fuss.  This type represents paycheck deductions and other automatically paid things that are expected to stop if the associated paycheck also stops.  The logic is that you do not save for these items, you simply never consider the money in your hand to begin with.