Fixes, but continued brokenness
It's barfing on transaction lines with comments, but debugging indicates the active transaction is null, so it isn't trying to parse the line as a transaction line.  Why in the world would that affect if the active transaction is null or not?  Not trusting debugger, but no idea what else it could be.