Fix for bad LedgerItem comparison
This showed up as an inability to reconcile transactions.  Making transactions conform to the immutable data class paradigm caused the rest of the program to no longer correctly consider the 'same' transactions as equivalent, which broke logic that makes such comparisons.  Now that LI has a unique persistent id, those kinds of issues should disappear.